+++ src/main/java/com/example/real_chat/api/command/RoomCommandApiController.java
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping(value = "api/v1/room/command")
+@RequestMapping(value = "api/v1/command/room")
 public class RoomCommandApiController {
 
     private final RoomCommandService roomService;
